Fix Context Menu Items
